+++ test/Unit/FakeVariablesTest.php
@@ -25,7 +25,7 @@ use PHPUnit\Framework;
  *
  * @uses \Ergebnis\Environment\Exception\InvalidName
  * @uses \Ergebnis\Environment\Exception\InvalidValue
- * @uses \Ergebnis\Environment\Test
+ * @uses \Ergebnis\Environment\TestVariables
  */
 final class FakeVariablesTest extends Framework\TestCase
 {